Added touch functionality
This adds touch functionality. The touchstart and onmouseover appear to be running at the same time. Functionality works fine with mouseover inputs, but touch devices do both functions. Results in undesired 3rd (reverse) animation state triggering  when the screen is tapped outside of the canvas area. (To be clear, I am strangely adding a different types of input; adding a tap on top of a hover. If designing a similar animation in the future, I would start with a click/tap because they are more similar.)

The whole code is a bit wacky, but still open to contributors pointing out how I can make it much more elegant, as I am sure there must be some very efficient shortcuts. Thanks!